requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.19
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
openpyxl==3.1.2
python-docx==0.8.11
deep-translator==1.11.4
python-jose==3.4.0
aiofiles==23.2.1
reportlab==3.6.12 